Fix: Extract RuntimeState to shared internal header
Problem:
RuntimeState struct was duplicated in two source files:
- hipdnn_ep_runtime_state.cpp (13 fields, complete definition)
- hipdnn_ep_runtime_tensor.cpp (was missing 6 fields, causing UB)

This violated the One Definition Rule (ODR) and caused undefined behavior
when RuntimeState objects crossed compilation unit boundaries.

Solution:
Created runtime_state_internal.h as the single source of truth for the
RuntimeState struct definition. Both implementation files now include
this shared header instead of defining the struct locally.

Benefits:
- Eliminates ODR violation
- Prevents struct definition drift
- Makes maintenance easier (single definition to update)
- Follows opaque pointer pattern correctly (public API only has forward
  declaration, implementation uses internal header)

Files changed:
- NEW: lib/Runtime/runtime_state_internal.h (RuntimeState definition)
- MODIFIED: lib/Runtime/hipdnn_ep_runtime_state.cpp (include internal header)
- MODIFIED: lib/Runtime/hipdnn_ep_runtime_tensor.cpp (include internal header)

Addresses review comment by fhanuman on PR-44.
