+++ lib/Runtime/runtime_state_internal.h
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+//===- runtime_state_internal.h - RuntimeState Internal Definition -------===//
+//
+// INTERNAL HEADER - Not part of public API
+//
+// This header defines the internal structure of RuntimeState.
+// Only runtime implementation files should include this header.
+//
+// Public interface uses opaque pointer (see hipdnn_ep_runtime.h).
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef HIPDNN_EP_RUNTIME_STATE_INTERNAL_H
+#define HIPDNN_EP_RUNTIME_STATE_INTERNAL_H
+
+#include "runtime_types.h"
+
+// Internal runtime state structure
+// This struct is opaque to generated code (passed as void*)
+struct RuntimeState {
+  hipStream_t stream;
+  miopenHandle_t miopen_handle;
+  hipblasLtHandle_t hipblas_handle;
+
+  // Single allocation holding all constants as one blob.
+  // gpu_constants[i] points into gpu_constants_blob at the offset stored in
+  // ConstantInfo, so only one allocation/copy is needed at init time.
+  // On dGPU: hipMalloc (VRAM). On iGPU: hipHostMalloc (pinned system RAM,
+  // GPU reads in-place, no hipMemcpy needed).
+  void *gpu_constants_blob;
+  bool constants_blob_is_host; // true = hipHostMalloc, false = hipMalloc
+  void **gpu_constants;
+  size_t num_constants;
+
+  // Memory pooling support
+  void *pool_base;        // Single large memory pool
+  size_t pool_size;       // Total pool size in bytes
+  size_t *buffer_offsets; // Offset for each buffer in the pool
+  size_t num_buffers;     // Number of buffers in the pool
+};
+
+#endif // HIPDNN_EP_RUNTIME_STATE_INTERNAL_H
